refactor momentum shift detector systematics
This allows modification of the reco muon momentum by an arbitary
function to generate a set of random systematics universes on the fly.

We generate a set of parameters for N universes, then create an array of
reco muon momenta for each event in each universe. The bin selection
TTreeFormulas are modified to compare to this array rather than the CV,
so that we get the reco bin set for each event/universe all at once.